+++ src/main/java/si/matjazcerkvenik/alertmonitor/util/OnStart.java
@@ -1,0 +1,22 @@
+package si.matjazcerkvenik.alertmonitor.util;
+
+import si.matjazcerkvenik.alertmonitor.model.DAO;
+
+import javax.servlet.ServletContextEvent;
+import javax.servlet.ServletContextListener;
+
+public class OnStart implements ServletContextListener {
+
+
+    @Override
+    public void contextInitialized(ServletContextEvent servletContextEvent) {
+
+        DAO.startUpTime = System.currentTimeMillis();
+
+    }
+
+    @Override
+    public void contextDestroyed(ServletContextEvent servletContextEvent) {
+        // no implementation
+    }
+}